+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+-   `Nullable`: Added new validator.
+-   `Optional`: Added new validator.
+
 ## [0.0.5] - 2022-01-14
 
 ### Changed